Class bloat refactoring -- move AlignmentTrack popup menu to top level class (#1227)
